clean(ZMSKVR): Remove changelog helper code duplication by introducing BaseChangelogHelper in zmsslim. Both zmsadmin and zmsstatistic now extend this base class.
